+++ python/tests/unit/eyes_robotframework/test_eyes_library.py
@@ -1,10 +1,12 @@
 import contextlib
 import os
+import sys
 from os.path import join
 from typing import Callable, Text
 
 import mock
 import pytest
+import yaml
 from AppiumLibrary import AppiumLibrary
 from SeleniumLibrary import SeleniumLibrary
 
@@ -83,3 +85,47 @@ def test_parse_config_from_env_variable(tmp_path):
     assert lib.selected_runner == SelectedRunner.web
     assert lib.config_path == "applitools.yaml"
     assert lib.configure
+
+
+@pytest.mark.parametrize(
+    "runner",
+    [
+        SelectedRunner.web.value,
+        SelectedRunner.web_ufg.value,
+        SelectedRunner.mobile_native.value,
+        SelectedRunner.native_mobile_grid.value,
+    ],
+)
+@pytest.mark.skipif(sys.version_info[0] == 2, reason="requires python3+")
+def test_parse_config_with_missing_runner_section(tmp_path, runner):
+    config_path = join(str(tmp_path), "applitools.yaml")
+    with eyes_lib_patcher(tmp_path):
+        with open(config_path, "r") as f:
+            raw_config = yaml.safe_load(f.read())
+
+        with open(config_path, "w") as f:
+            raw_config.pop(runner)
+            yaml.dump(raw_config, f)
+        with pytest.raises(KeyError):
+            lib = EyesLibrary(config=config_path, runner=runner)
+
+
+@pytest.mark.parametrize(
+    "runner",
+    [
+        SelectedRunner.web_ufg.value,
+        SelectedRunner.mobile_native.value,
+        SelectedRunner.native_mobile_grid.value,
+    ],
+)
+@pytest.mark.skipif(sys.version_info[0] == 2, reason="requires python3+")
+def test_parse_config_with_missing_sections(tmp_path, runner):
+    config_path = join(str(tmp_path), "applitools.yaml")
+    with eyes_lib_patcher(tmp_path):
+        with open(config_path, "r") as f:
+            raw_config = yaml.safe_load(f.read())
+
+        with open(config_path, "w") as f:
+            raw_config.pop(runner)
+            yaml.safe_dump(raw_config, f)
+        lib = EyesLibrary(config=config_path)