+++ python/eyes_robotframework/CHANGELOG.md
@@ -1,3 +1,7 @@
+[vNext]
+### Fixed
+- Not possible to remove runner related sections from config [Trello 3016](https://trello.com/c/B0V3HPtF)
+
 [5.8.0] - 2022-07-26
 ### Added
 - Coded region keywords now support optional padding and region_id arguments [Trello 2977](https://trello.com/c/lYJnhwCz)